requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.2
fastapi-utils==0.2.1
filelock==3.13.4
fsspec==2024.3.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
huggingface-hub==0.22.2
idna==3.7
iniconfig==2.0.0
joblib==1.4.0
numpy==1.24.4
packaging==24.0
pkg_resources==0.0.0
pluggy==1.5.0
pydantic==1.10.15
pydantic_core==2.18.2
pytest==8.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
skops==0.9.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.37.2
tabulate==0.9.0
threadpoolctl==3.4.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
